This is the initial version of getbrother.php.
I made it to primarily use it in Zabbix, so copy it over to the location of external scripts for Zabbix, make it executable, and call it as an external check within the templates:

For black toner:
getbrother.php[{HOST.IP},81]

For cyan toner:
getbrother.php[{HOST.IP},82]

For magenta toner:
getbrother.php[{HOST.IP},83]

For yellow toner:
getbrother.php[{HOST.IP},84]

It can also be used in Nagios or other infrastructure monitoring tools on Linux.  It might also work in Windows, but this wasn't developed there.

You can also call it stand-alone if it's executable:

./getbrother.php <ip> <code>

Example:
  ./getbrother.php 192.168.100.106 81